Use externalpipe to follow urls
Also update the Makefile to install and uninstall the helper
externalpipe scripts to the same directory as the st executable. This
way the scripts will also be in the path so that st can call them.

When developing new scripts, use local paths instead - the local st
executable is unable to find the scripts in the scripts directory. An
alternative would be to add the local directory to $PATH for the
duration of testing.