Change dmenu_nice invocation back to dmenu
The dmenu_nice script was written as a wrapper around dmenu because
after dmenu 4.9 was run, the active window would not regain focus. This
script would keep track of which window had focus, call dmenu, and
return focus to the active window.

The recommendation to fix this issue is to downgrade to dmenu 4.8. I
have now forked this version of dmenu rendering the dmenu_nice script
pretty much useless. Hence the reference in st_copy_url is being
reverted to dmenu.